chore: organized some components
