--- src/app/[locale]/courses/[id]/goals/page.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function CourseGoals() {
-  return <div>Goals</div>;
-}
