+++ src/app/[locale]/courses/page.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 import styles from './courses.module.css';
-import Course from '@/components/course';
+import Course from '@/components/course/course';
 import data from '../../mockdata.json';
 
 export default async function CourseHome() {
