+++ src/app/[locale]/reading/page.tsx
@@ -1,11 +1,7 @@
-import Image from 'next/image';
-import { remark } from 'remark';
-import html from 'remark-html';
 import { cookies } from 'next/headers';
 import { createServerComponentClient } from '@supabase/auth-helpers-nextjs';
 import { Database } from '@/lib/database.types';
 import ArticleCard from '@/components/ArticleCard';
-import Link from 'next/link';
 
 export default async function Reading() {
   // Fetch title, description, id and image path of articles from superbase database
