+++ src/components/course/contentSection.tsx
